+++ components/PButtonTypo/PButtonTypo.styled.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export const StyledButtonTypo = styled.p`
+  font-size: 0.625rem;
+  font-weight: 400;
+  color: #ffffff;
+  text-align: center;
+  margin: 0rem 0rem 0.1rem 0.2rem;
+`;
